Functionality to add and delete approvals added.

+++ dcbackend/dcertificate/issuer.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, g, request
+import sqlite3
 from dcertificate.db import get_db
 from dcertificate.auth import require_issuer_login
 import dcertificate.lib as lib
@@ -25,7 +26,7 @@ def certifications_list():
         "data": certifications
     }, 200
 
-@bp.get("approvals-list")
+@bp.get("/approval/list")
 @require_issuer_login
 def approvals_list():
     issuer_id = g.issuer_id
@@ -52,13 +53,33 @@ def approvals_list():
         "data": approvals
     }, 200
 
-@bp.delete("/delete-approval")
+@bp.delete("/approval/delete")
 @require_issuer_login
 def delete_approval():
     issuer_id = g.issuer_id
     request_json = request.json
+
+    if 'certification_id' not in request_json:
+        return {
+            'success': False,
+            'debug': 'certification_id not in request json.'
+        }, 400
+    
     db = get_db()
 
+    query_response = db.execute(
+        "SELECT * FROM approval "
+        "WHERE issuer_id = ? "
+        "AND certification_id = ?;",
+        (issuer_id, request_json['certification_id'])
+    ).fetchall()
+
+    if(len(query_response) == 0):
+        return {
+            'success': False,
+            'message': 'Certification with given id does not exist.'
+        }, 404
+    
     query_response = db.execute(
         "DELETE FROM approval "
         "WHERE issuer_id = ? "
@@ -72,3 +93,83 @@ def delete_approval():
         "success": True,
         "message": "Approval deleted."
     }, 200
+
+@bp.get("/approval/certification/<certification_id>")
+@require_issuer_login
+def certification_for_approval(certification_id:int):
+    issuer_id = g.issuer_id
+
+    try:
+        certification_id = int(certification_id)
+    except TypeError:
+        return {
+            'success': False,
+            'message': 'Invalid certification ID.',
+            'debug': 'certification_id must be int.'
+        }, 400
+    
+    db = get_db()
+
+    query_response = db.execute(
+        'SELECT certification.id AS id, '
+        'certification.title AS title, '
+        'issuer.display_name AS issuer_display_name ,'
+        'certification.issuer_id AS issuer_id '
+        'FROM certification '
+        'LEFT JOIN issuer ON issuer.id = certification.issuer_id '
+        'WHERE certification.id = ?;',
+        (certification_id,)
+    ).fetchall()
+
+    if(len(query_response) == 0):
+        return {
+            'success': False,
+            'message': 'Certification not found.'
+        }, 404
+    
+    data = dict(query_response[0])
+        
+    return {
+        'success': True,
+        'data': data
+    }, 200
+    
+@bp.delete("/approval/add")
+@require_issuer_login
+def add_approval():
+    issuer_id = g.issuer_id
+    try:
+        certification_id = int(request.json['certification_id'])
+    except KeyError:
+        return {
+            'success': False,
+            'debug': 'certification_id not found in requested json.'
+        }, 400
+    except TypeError:
+        return {
+            'success': False,
+            'message': 'Invalid certification ID.',
+            'debug': 'certification_id must be int.'
+        }, 400
+    else:
+    
+        db = get_db()
+        try:
+            db.execute(
+                'INSERT INTO approval '
+                '(issuer_id, certification_id) '
+                'VALUES (?,?);',
+                (issuer_id, certification_id)
+            )
+        except sqlite3.IntegrityError:
+            return {
+                'success': False,
+                'message': 'Approval already exists.'
+            }
+        else:
+            db.commit()
+            
+            return {
+                'success': True,
+                'message': 'Approval added successfully.'
+            }
